tweaks to offer cell and removed viewModel

+++ ChurnSmarter/Models/Credit.swift
@@ -2,6 +2,6 @@ import Foundation
 
 struct Credit: Decodable, Hashable {
     let description: String?
-    let value: Double?
+    let value: Int?
     let weight: Double?
 }
